src/requirements.txt
Django==4.0.7
django-celery-beat==2.3.0
djangorestframework==3.13.1
djoser==2.1.0
celery==5.2.7
redis==4.3.4
psycopg2-binary==2.9.3
python-decouple==3.6